requirements.txt
acme==1.1.0
alabaster==0.7.8
appdirs==1.4.3
asgiref==3.5.2
attrs==19.3.0
Babel==2.6.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
blinker==1.4
cached-property==1.5.1
cachetools==5.2.0
certbot==0.40.0
certbot-nginx==0.40.0
certifi==2022.6.15
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.0
Click==7.0
colorama==0.4.3
ConfigArgParse==0.13.0
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
datefinder==0.7.1
dbus-python==1.2.16
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
Django==4.0.4
django-cors-headers==3.13.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt
djoser==2.1.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.16
drf-yasg==1.21.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
flake8==4.0.1
flake8-ugettext-alias==1.1
future==0.18.2
gitsome==0.8.0
google-api-core==2.8.2
google-api-python-client==2.53.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
heroku==0.1.4
httplib2==0.20.4
idna==3.3
imagesize==1.2.0
importlib-metadata==1.5.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
josepy==1.2.0
jsonschema==3.2.0
keyring==18.0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==2.1.1
mccabe==0.6.1
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
numpydoc==0.7.0
oauth2client==1.5.2
oauthlib==3.2.0
olefile==0.46
openapi-codec==1.3.2
packaging==21.3
paramiko==2.6.0
parsedatetime==2.4
pbr==5.4.5
pexpect==4.6.0
Pillow==7.0.0
ply==3.11
prompt-toolkit==2.0.10
protobuf==4.21.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.21
pycups==1.9.73
pyflakes==2.4.0
Pygments==2.3.1
PyGObject==3.36.0
PyICU==2.4.2
PyJWT==2.4.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.15.5
python-dateutil==1.5
python3-openid==3.2.0
pytz==2022.1
pyudev==0.21.0
pyxdg==0.26
PyYAML==5.3.1
regex==2022.7.9
reportlab==3.5.34
requests==2.28.1
requests-oauthlib==1.3.1
requests-toolbelt==0.8.0
requests-unixsocket==0.2.0
roman==2.0.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==2.3.1
setproctitle==1.1.10
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
Sphinx==1.8.5
sqlparse==0.4.2
supervisor==4.1.0
systemd-python==234
testresources==2.0.1
texttable==1.6.2

uritemplate==4.1.1
urllib3==1.26.10
virtualenv==20.0.17
wadllib==1.3.3
wcwidth==0.1.8
websocket-client==0.53.0
whitenoise==0.13
xonsh==0.9.13
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.0.0
zope.interface==4.7.1
